+++ packages/schema/lts.ts
@@ -1,3 +1,11 @@
 /** This is the source of truth for currently supported Schema Versions by this library */
 
-module.exports = ["1.5.1", "1.6.0", "1.7.0", "1.8.0", "1.9.0", "1.10.0"];
+module.exports = [
+  "1.5.1",
+  "1.6.0",
+  "1.7.0",
+  "1.8.0",
+  "1.9.0",
+  "1.10.0",
+  "1.11.0",
+];