+++ packages/schema/src/index.ts
@@ -2,7 +2,15 @@
 export * as lts from "./lts";
 
 /* Always re-export the latest version directly */
-export * from "./lts/1.10.0";
+export * from "./lts/1.11.0";
 
 /* And finally export a list of LTS versions */
-export const LTS_VERSIONS = ["1.5.1","1.6.0","1.7.0","1.8.0","1.9.0","1.10.0"];
+export const LTS_VERSIONS = [
+  "1.5.1",
+  "1.6.0",
+  "1.7.0",
+  "1.8.0",
+  "1.9.0",
+  "1.10.0",
+  "1.11.0",
+];