+++ packages/belt/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @versaprotocol/belt
 
+## 1.10.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @versaprotocol/schema@1.11.0
+
 ## 1.10.0
 
 ### Minor Changes